chore: shift CTA from bin to .ts
