+++ tooling/create-tauri-app/.eslintrc.js
@@ -7,7 +7,7 @@ module.exports = {
   },
 
   parser: '@typescript-eslint/parser',
-  ignorePatterns: ['.eslintrc.js', 'jest.config.js', 'test/**/*'],
+  ignorePatterns: ['.eslintrc.js', 'jest.config.js', 'test', 'bin'],
   extends: [
     'standard-with-typescript',
     'plugin:@typescript-eslint/recommended-requiring-type-checking',
