+++ tooling/create-tauri-app/package.json
@@ -45,6 +45,7 @@
     "@types/cross-spawn": "6.0.2",
     "@types/inquirer": "7.3.1",
     "@types/jest": "26.0.23",
+    "@types/minimist": "1.2.1",
     "@types/semver": "7.3.5",
     "@typescript-eslint/eslint-plugin": "4.22.0",
     "@typescript-eslint/parser": "4.22.0",
